+++ src/components/Movie/MovieMain/MovieMain.tsx
@@ -1,6 +1,11 @@
 import { Container } from '../../index';
 import s from './MovieMain.module.scss';
+import { useEffect, useState } from 'react';
 import { MovieButtonWishList } from '../MovieButtonWishList/MovieButtonWishList';
+import { useAppSelector, useAppDispatch } from '../../../hooks/hooks';
+import { pushMovie } from '../../../store/wishlistSlice';
+import { pushMovieInUser } from '../../../store/userSlice';
+import { getTokenUser } from '../../../helpers/getTokenUser';
 
 import imdbImage from '../../../assets/imdb.svg';
 
@@ -12,6 +17,28 @@ type MovieMainProps = {
 };
 
 export const MovieMain = ({ img, title, imdb, id }: MovieMainProps) => {
+ const [dataUsers, setDataUsers] = useState([]);
+ const user = useAppSelector((state) => state.user.user);
+ const movies = useAppSelector((state) => state.movies.movies);
+ const wishlist = useAppSelector((state) => state.wishlist.wishlist);
+ const dispatch = useAppDispatch();
+
+ useEffect(() => {
+  fetch('https://aston-moviebox-default-rtdb.firebaseio.com/users.json')
+   .then((response) => response.json())
+   .then((json) => setDataUsers(json));
+ }, []);
+
+ const addInWishlist = () => {
+  const movie = movies.find((item) => item.id === +id);
+  const moviesArray = [...wishlist, movie];
+  const inWishlist = wishlist.find((item) => item.id === +movie.id);
+  if (!inWishlist) {
+   const token = getTokenUser(dataUsers, user.mail);
+   dispatch(pushMovie(movie));
+   dispatch(pushMovieInUser({ token, moviesArray }));
+  }
+ };
  return (
   <div className={s.movie}>
    <img src={img} alt={img} className={s.movie_main} />
@@ -27,7 +54,11 @@ export const MovieMain = ({ img, title, imdb, id }: MovieMainProps) => {
         </div>
        </div>
       )}
-      <MovieButtonWishList id={id}>Add to WishList</MovieButtonWishList>
+      {user.name && (
+       <MovieButtonWishList id={id} click={addInWishlist}>
+        Add to WishList
+       </MovieButtonWishList>
+      )}
      </div>
     </Container>
    </div>
